+++ fastestserver/src/types.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 pub struct LevelResponse {
     pub level: i32,
-    pub time: i32,
+    pub time: i64,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -17,7 +17,7 @@ pub struct AllLevelResponse {
 pub struct RankResponse {
     pub rank: i32,
     pub rank_name: String,
-    pub time: i32,
+    pub time: i64,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -29,7 +29,7 @@ pub struct AllRankResponse {
 pub struct InsertRequest {
     pub in_game_name: String,
     pub r#type: String,
-    pub time: i32,
+    pub time: i64,
     pub game_name: String,
 }
 
@@ -38,7 +38,7 @@ pub struct InsertLevelRequest {
     pub in_game_name: String,
     pub old_level: i32,
     pub new_level: i32,
-    pub time: i32,
+    pub time: i64,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -48,7 +48,7 @@ pub struct InsertRankRequest {
     pub old_rank_name: String,
     pub new_rank: i32,
     pub new_rank_name: String,
-    pub time: i32,
+    pub time: i64,
     pub rank_type: String,
 }
